+++ app/src/main/java/com/poupa/vinylmusicplayer/loader/PlaylistSongLoader.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 public class PlaylistSongLoader {
 
     @NonNull
-    public static ArrayList<Song> getPlaylistSongList(@NonNull final Context context, final int playlistId) {
+    public static ArrayList<Song> getPlaylistSongList(@NonNull final Context context, final long playlistId) {
         ArrayList<Song> songs = new ArrayList<>();
         Cursor cursor = makePlaylistSongCursor(context, playlistId);
 
@@ -32,8 +32,8 @@ public class PlaylistSongLoader {
     }
 
     @NonNull
-    private static PlaylistSong getPlaylistSongFromCursorImpl(@NonNull Cursor cursor, int playlistId) {
-        final int id = cursor.getInt(0);
+    private static PlaylistSong getPlaylistSongFromCursorImpl(@NonNull Cursor cursor, long playlistId) {
+        final long id = cursor.getLong(0);
         final String data = cursor.getString(5);
         final int dateAdded = cursor.getInt(6);
         final int dateModified = cursor.getInt(7);
@@ -67,9 +67,9 @@ public class PlaylistSongLoader {
         final int trackNumber = cursor.getInt(2);
         final int year = cursor.getInt(3);
         final long duration = cursor.getLong(4);
-        final int albumId = cursor.getInt(8);
+        final long albumId = cursor.getLong(8);
         final String albumName = cursor.getString(9);
-        final int artistId = cursor.getInt(10);
+        final long artistId = cursor.getLong(10);
         final String artistName = cursor.getString(11);
 
         PlaylistSong playlistSong = new PlaylistSong(
@@ -92,7 +92,7 @@ public class PlaylistSongLoader {
         return playlistSong;
     }
 
-    public static Cursor makePlaylistSongCursor(@NonNull final Context context, final int playlistId) {
+    public static Cursor makePlaylistSongCursor(@NonNull final Context context, final long playlistId) {
         try {
             return context.getContentResolver().query(
                     MediaStore.Audio.Playlists.Members.getContentUri("external", playlistId),
