+++ app/src/main/java/com/poupa/vinylmusicplayer/model/AbsCustomPlaylist.java
@@ -7,7 +7,7 @@ import android.os.Parcel;
  */
 
 public abstract class AbsCustomPlaylist extends Playlist {
-    public AbsCustomPlaylist(int id, String name) {
+    public AbsCustomPlaylist(long id, String name) {
         super(id, name);
     }
