+++ app/src/main/java/com/poupa/vinylmusicplayer/model/Song.java
@@ -11,7 +11,7 @@ import com.poupa.vinylmusicplayer.loader.ReplayGainTagExtractor;
 public class Song implements Parcelable {
     public static final Song EMPTY_SONG = new Song(-1, "", -1, -1, -1, "", -1, -1, -1, "", -1, "");
 
-    public final int id;
+    public final long id;
     public String title;
     public int trackNumber;
     public int year;
@@ -19,15 +19,15 @@ public class Song implements Parcelable {
     public final String data;
     public final long dateAdded;
     public final long dateModified;
-    public final int albumId;
+    public final long albumId;
     public String albumName;
-    public final int artistId;
+    public final long artistId;
     public String artistName;
 
     private float replaygainTrack = Float.NaN;
     private float replaygainAlbum = Float.NaN;
 
-    public Song(int id, String title, int trackNumber, int year, long duration, String data, long dateAdded, long dateModified, int albumId, String albumName, int artistId, String artistName) {
+    public Song(long id, String title, int trackNumber, int year, long duration, String data, long dateAdded, long dateModified, long albumId, String albumName, long artistId, String artistName) {
         this.id = id;
         this.title = title;
         this.trackNumber = trackNumber;
@@ -88,7 +88,7 @@ public class Song implements Parcelable {
 
     @Override
     public int hashCode() {
-        int result = id;
+        int result = (int)id;
         result = 31 * result + (title != null ? title.hashCode() : 0);
         result = 31 * result + trackNumber;
         result = 31 * result + year;
@@ -96,9 +96,9 @@ public class Song implements Parcelable {
         result = 31 * result + (data != null ? data.hashCode() : 0);
         result = 31 * result + (int) (dateAdded ^ (dateAdded >>> 32));
         result = 31 * result + (int) (dateModified ^ (dateModified >>> 32));
-        result = 31 * result + albumId;
+        result = 31 * result + (int)albumId;
         result = 31 * result + (albumName != null ? albumName.hashCode() : 0);
-        result = 31 * result + artistId;
+        result = 31 * result + (int)artistId;
         result = 31 * result + (artistName != null ? artistName.hashCode() : 0);
         return result;
     }
@@ -128,7 +128,7 @@ public class Song implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeString(this.title);
         dest.writeInt(this.trackNumber);
         dest.writeInt(this.year);
@@ -136,14 +136,14 @@ public class Song implements Parcelable {
         dest.writeString(this.data);
         dest.writeLong(this.dateAdded);
         dest.writeLong(this.dateModified);
-        dest.writeInt(this.albumId);
+        dest.writeLong(this.albumId);
         dest.writeString(this.albumName);
-        dest.writeInt(this.artistId);
+        dest.writeLong(this.artistId);
         dest.writeString(this.artistName);
     }
 
     protected Song(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.title = in.readString();
         this.trackNumber = in.readInt();
         this.year = in.readInt();
@@ -151,9 +151,9 @@ public class Song implements Parcelable {
         this.data = in.readString();
         this.dateAdded = in.readLong();
         this.dateModified = in.readLong();
-        this.albumId = in.readInt();
+        this.albumId = in.readLong();
         this.albumName = in.readString();
-        this.artistId = in.readInt();
+        this.artistId = in.readLong();
         this.artistName = in.readString();
     }
