+++ app/src/main/java/com/poupa/vinylmusicplayer/service/MusicService.java
@@ -1028,7 +1028,7 @@ public class MusicService extends MediaBrowserServiceCompat implements SharedPre
                 break;
             case SHUFFLE_MODE_NONE:
                 this.shuffleMode = shuffleMode;
-                int currentSongId = getCurrentSong().id;
+                long currentSongId = getCurrentSong().id;
                 playingQueue = new ArrayList<>(originalPlayingQueue);
                 int newPosition = 0;
                 for (Song song : getPlayingQueue()) {
